requirements.txt
faker
psycopg2-binary
pytz
logging
sqlalchemy
pandas
python-dotenv